// src/services/model/model/smsHomeBrand.ts
// 此文件由 oneapi-cli 自动生成，请不要编辑（This file is auto-generated by oneapi-cli, Please do not edit it）
/** */
export interface SmsHomeBrand {
  /** */
  id: number;
  /** */
  brandId: number;
  /** */
  brandName: string;
  /** */
  recommendStatus: number;
  /** */
  sort: number;
}